updating logger to simple-logger
